% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolcano.R
\name{deconvolcano}
\alias{deconvolcano}
\title{Make a deconvolcano plot}
\usage{
deconvolcano(
  normdata,
  metadata,
  feature,
  lab.1,
  lab.2,
  ID,
  n_cell_proxy,
  neg_probe_names
)
}
\arguments{
\item{normdata}{A normalized and quality-controlled counts table with the
column names being Sample IDs or Cell IDs and the row names being gene IDs.
Class: data frame.}

\item{metadata}{A meta data table that has an ID column with identical IDs to
the column names of normdata. Class: data frame.}

\item{feature}{The name of the column in metadata that contains the two
groups for which differential expression should be performed. Class: string.}

\item{lab.1}{A label in the feature column that is to be compared to lab.2.
Class: string. Note: needs to be before lab.2 alphabetically.}

\item{lab.2}{A label in the feature column that is to be compared to lab.1.
Class: string. Note: needs to be after lab.1 alphabetically.}

\item{ID}{The name of the ID column in metadata. Class: string.}

\item{n_cell_proxy}{The name of the column in metadata that describes a
variable that can be used as a proxy for the number of cells in each AOI.
Often this is 'AOINucleiCount' but not always. Class: string.}

\item{neg_probe_names}{The names of the negative probes as they appear in the
row names of normdata. Class: vector.}
}
\description{
This function prints a volcano plot that depicts differential containment of
immune cell types across two groups of AOIs in Nanostring GeoMx data.
}
