% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.expression.R
\name{differential.expression}
\alias{differential.expression}
\title{Perform differential expression analysis}
\usage{
differential.expression(
  counts.table,
  meta.data,
  meta.feature,
  value.1,
  value.2,
  test = "wilcox"
)
}
\arguments{
\item{counts.table}{A normalized and quality controlled counts table. Class:
data frame.}

\item{meta.data}{A data frame of sample names and groups to perform
differential expression on. The row names must exist as column names of
counts.table.}

\item{meta.feature}{The name of the column in meta.data for which the groups
being compared exist.}

\item{value.1}{One of the two labeled groups being compared. This will be a
level of the meta.feature column.}

\item{value.2}{The second of the labeled groups being compared. This will be
a level of the meta.feature column.}

\item{test}{The statistical test that will be run to compare the two groups.
This test must be "wilcox" or "t" and nothing else.}
}
\description{
This function returns a table with differential expression data for genes in
a counts table.
}
