% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy.gsea.R
\name{easy.gsea}
\alias{easy.gsea}
\title{Do easy GSEA}
\usage{
easy.gsea(de.table, order.stat, ontology, organism)
}
\arguments{
\item{de.table}{The output of the diff.exp function. A table containing
differential expression results.}

\item{order.stat}{The statistic in de.table that will be used to order the
differentially expressed genes. Typically, one will use "log2.FC" as this
parameter's value.}

\item{ontology}{Either "BP", "MP", or "CC". BP is for biological
processes, MP is for molecular function, and CC is for cellular component.}

\item{organism}{Either "human" or "mouse". Based on the value, the function
will reference the corresponding database of genes.}
}
\description{
This function is a wrapper for the gseGO function of the clusterProfiler
package. It implements the function in a simpler way.
}
