% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyGSEA.R
\name{EasyGSEA}
\alias{EasyGSEA}
\title{Easily perform GSEA}
\usage{
EasyGSEA(
  normdata,
  metadata,
  directory,
  feature,
  lab.1,
  lab.2,
  ID,
  gene_set,
  num_perm
)
}
\arguments{
\item{normdata}{A normalized and quality-controlled counts table with the
first column being Gene IDs. Class: data frame.}

\item{metadata}{A meta data table that has an ID column with identical IDs to
the column names of normdata. Class: data frame.}

\item{directory}{The working directory to send output files to. Class: string.}

\item{feature}{The name of the column in metadata that contains the two
groups for which differential expression should be performed. Class: string.}

\item{lab.1}{One label in the feature column that is to be compared to lab.2.
Class: string.}

\item{lab.2}{One label in the feature column that is to be compared to lab.1.
Class: string.}

\item{ID}{The name of the ID column in metadata. Class: string.}

\item{gene_set}{The file path to a gene set in .gmt format. This file is
downloadable from the GSEA website. Class: string.}

\item{num_perm}{The number of permutations used to obtain the p value. 1000
is recommended. Class = integer.}
}
\description{
This function performs GSEA without the headache of getting files in specific
formats that the original GSEA function requires. It also eliminates the many
parameters the original GSEA function requires by simply setting them as the
values that will be most often used by the Ting Lab. One should revert to the
original GSEA function, if trying to perform GSEA with atypical parameters.
}
