% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diffexp.R
\name{get_diffexp}
\alias{get_diffexp}
\title{Get differential expression}
\usage{
get_diffexp(df, de)
}
\arguments{
\item{df}{A normalized and quality controlled counts table. Class: data frame.}

\item{de}{A data frame of sample names and groups to perform DE on.}
}
\description{
This function returns a table with differential expression data for genes in
a counts table.
}
