% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.plot.R
\name{volcano.plot}
\alias{volcano.plot}
\title{Make a volcano plot to show differential expression}
\usage{
volcano.plot(
  diff.exp.results,
  color.vals = c("blue", "black", "red"),
  log2.FC.cutoff = 1,
  p.adj.cutoff = 0.05,
  label.res = 50
)
}
\arguments{
\item{diff.exp.results}{The data frame that is outputted by the diff.exp
function.}

\item{color.vals}{A vector of three color values. The first color will
correspond to genes up-regulated in the first group being compared in one's
differential expression analysis. The second color will correspond to genes
up-regulated in neither group. The third color will correspond to genes
up-regulated in the third group.}

\item{log2.FC.cutoff}{The LFC cutoff for genes to be labeled on the plot.}

\item{p.adj.cutoff}{The p.adj cutoff for genes to be labeled on the plot.}

\item{label.res}{Higher values mean more overlaps are allowed for gene
labeling purposes on the plot.}
}
\description{
This function can be used to make a volcano plot showing which genes are
differentially expressed across two groups of AOIs or cells. Differential
expression is calculated using a wilcox rank sum test or a student's t test
for difference in mean expression across the two groups.
}
