% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano}
\alias{volcano}
\title{Make a volcano plot to show differential expression}
\usage{
volcano(normdata, metadata, feature, lab.1, lab.2, ID)
}
\arguments{
\item{normdata}{A normalized and quality-controlled counts table with the
column names being Sample IDs or Cell IDs and the row names being gene IDs.
Class: data frame.}

\item{metadata}{A meta data table that has an ID column with identical IDs to
the column names of normdata. Class: data frame.}

\item{feature}{The name of the column in metadata that contains the two
groups for which differential expression should be performed. Class: string.}

\item{lab.1}{A label in the feature column that is to be compared to lab.2.
Class: string. Note: needs to be before lab.2 alphabetically.}

\item{lab.2}{A label in the feature column that is to be compared to lab.1.
Class: string. Note: needs to be after lab.1 alphabetically.}

\item{ID}{The name of the ID column in metadata. Class: string.}
}
\description{
This function can be used to make a volcano plot showing which genes are
differentially expressed across two groups of AOIs or cells. Differential
expression is calculated using a wilcox rank sum test for difference in mean
expression across the two groups.
}
